Move `registerDependenciesTask` from SpotlessExtensionImpl to SpotlessExtension.
